Move FluentAttributeContainer into own package
